Handle queries for comment counts properly
A comment query can return either an array or an int result. As we
were only returning an empty array for any comment query, a fatal
error would be generated if someone requested a count.

This specifically happens when the `_embed` parameter is added to
a REST API call for posts.

Fixes #35